Time: 4 ms (55.15%), Space: 53.5 MB (5.27%) - LeetHub

+++ 0169-majority-element/0169-majority-element.java
@@ -4,6 +4,6 @@ class Solution {
         
         int n = nums.length;
         
-        return nums[ (int) n/2 ];
+        return nums[n/2];
     }
 }
